requirements.txt
flask==3.0.0
google-genai
pillow==10.1.0

